Rework util.go (removed) and work with directories (dirs.go)
This change gets rid of util.go that accumulated too much types
and functions that should live next to where they used. Also,
it introduces a centralised approach to managing setting and
creation of the directories used by workspaced
