+++ internal/overlord/hookstate/handlers.go
@@ -4,9 +4,8 @@ import (
 	"fmt"
 	"path/filepath"
 
-	util "github.com/canonical/workspace/internal"
-	. "github.com/canonical/workspace/internal/overlord/sharedstate"
 	"github.com/canonical/workspace/internal/overlord/state"
+	. "github.com/canonical/workspace/internal/overlord/sthelper"
 	"github.com/canonical/workspace/internal/workspacebackend"
 
 	"github.com/spf13/afero"
@@ -29,7 +28,7 @@ func (h *HookManager) doRunHook(task *state.Task, tomb *tomb.Tomb) error {
 
 	st := task.State()
 	st.Lock()
-	var hook util.WorkspaceHookType
+	var hook workspacebackend.WorkspaceHookType
 	err = task.Get("hook-setup", &hook)
 	st.Unlock()
 	if err != nil {
@@ -38,7 +37,7 @@ func (h *HookManager) doRunHook(task *state.Task, tomb *tomb.Tomb) error {
 
 	/* create a memory out/err to log the hook output into the task's log */
 	memFs := afero.NewMemMapFs()
-	outerr, err := memFs.Create(util.ToInstanceName(workspace, prj.ProjectId))
+	outerr, err := memFs.Create(workspacebackend.InstanceName(workspace, prj.ProjectId))
 	if err != nil {
 		return err
 	}
@@ -53,9 +52,9 @@ func (h *HookManager) doRunHook(task *state.Task, tomb *tomb.Tomb) error {
 			"-o",
 			"pipefail",
 			"-c",
-			filepath.Join(util.ToHooksPath(blob.Name), hook.String()),
+			filepath.Join(workspacebackend.SdkHooksPath(blob.Name), hook.String()),
 		},
-		WorkDir: util.ToHooksPath(blob.Name),
+		WorkDir: workspacebackend.SdkHooksPath(blob.Name),
 		Stdin:   nil,
 		Stdout:  outerr,
 		Stderr:  outerr}
