+++ internal/overlord/sdkstate/handlers.go
@@ -6,12 +6,13 @@ import (
 	"path/filepath"
 	"strconv"
 
-	util "github.com/canonical/workspace/internal"
+	"github.com/canonical/workspace/internal/dirs"
 	store "github.com/canonical/workspace/internal/fakestore"
 	"github.com/canonical/workspace/internal/logger"
-	. "github.com/canonical/workspace/internal/overlord/sharedstate"
 	"github.com/canonical/workspace/internal/overlord/state"
-	backend "github.com/canonical/workspace/internal/workspacebackend"
+	. "github.com/canonical/workspace/internal/overlord/sthelper"
+	"github.com/canonical/workspace/internal/workspacebackend"
+
 	"github.com/spf13/afero"
 
 	"gopkg.in/tomb.v2"
@@ -19,7 +20,7 @@ import (
 
 func (m *SdkManager) doRetrieveSdk(task *state.Task, tomb *tomb.Tomb) error {
 	st := task.State()
-	var sdk backend.Sdk
+	var sdk workspacebackend.Sdk
 
 	st.Lock()
 	err := task.Get("sdk-setup", &sdk)
@@ -34,7 +35,7 @@ func (m *SdkManager) doRetrieveSdk(task *state.Task, tomb *tomb.Tomb) error {
 		return nil
 	}
 
-	blob, err := client.RetrieveSdk(sdk.Name, sdk.Channel)
+	blob, err := client.RetrieveSdk(sdk.Name, sdk.Channel, dirs.SdkDir)
 	if err != nil {
 		return err
 	}
@@ -46,14 +47,12 @@ func (m *SdkManager) doRetrieveSdk(task *state.Task, tomb *tomb.Tomb) error {
 	return nil
 }
 
-func sdkBlobDevice(sdk *store.SdkBlob) backend.WorkspaceDevice {
-	filename := store.ToSdkFilename(sdk.Name, sdk.Revision)
-
+func sdkBlobDevice(sdk *store.SdkBlob) workspacebackend.WorkspaceDevice {
 	/* Bind-mount the SDK to the workspace */
-	return backend.WorkspaceDevice{
+	return workspacebackend.WorkspaceDevice{
 		Name: sdk.Name,
-		Properties: map[string]string{"type": "disk", "source": filename,
-			"path": filepath.Join("/root", filepath.Base(filename))},
+		Properties: map[string]string{"type": "disk", "source": sdk.Filename,
+			"path": filepath.Join("/root", filepath.Base(sdk.Filename))},
 	}
 }
 
@@ -92,19 +91,19 @@ func (m *SdkManager) doInstallSDK(task *state.Task, tomb *tomb.Tomb) error {
 	defer cleanup()
 
 	/* example: /var/lib/workspace/sdk/cuda/712/ */
-	sdkPath := filepath.Join(util.WorkspaceSdksDir, blob.Name,
+	sdkPath := filepath.Join(workspacebackend.WorkspaceSdksDir, blob.Name,
 		strconv.Itoa(int(blob.Revision)))
 
 	/* create a memory out/err to log the hook output into the task's log */
 	memFs := afero.NewMemMapFs()
-	outerr, err := memFs.Create(util.ToInstanceName(workspace, project.ProjectId))
+	outerr, err := memFs.Create(workspacebackend.InstanceName(workspace, project.ProjectId))
 	if err != nil {
 		return err
 	}
 
 	/* Unpack the SDK to the desired location in the workspace
 	   Note: the following command requires ~ tar >= 1.29 due to --one-top-level */
-	args := backend.ExecArgs{
+	args := workspacebackend.ExecArgs{
 		User: "root",
 		Command: []string{
 			"tar",
@@ -158,7 +157,7 @@ func (m *SdkManager) undoInstallSdk(task *state.Task, tomb *tomb.Tomb) error {
 	}
 	defer fs.Close()
 
-	err = fs.RemoveAll(filepath.Join(util.WorkspaceSdksDir, blob.Name))
+	err = fs.RemoveAll(filepath.Join(workspacebackend.WorkspaceSdksDir, blob.Name))
 	if err != nil {
 		return fmt.Errorf("cannot undo SDK %q installation: %w", sdkMount.Name, err)
 	}
@@ -208,7 +207,7 @@ func (m *SdkManager) doLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
 	/* Make a record in a LXD's key value storage to maintain
 	the sequence of the SDK's revisions */
 	err = m.backend.AddWorkspaceConfig(ctx, workspace,
-		&backend.WorkspaceConfigValue{
+		&workspacebackend.WorkspaceConfigValue{
 			Name:  "user.workspace.sdk",
 			Value: string(sequenceValue),
 		})
@@ -218,7 +217,7 @@ func (m *SdkManager) doLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
 	}
 
 	/* Update the current link to point out to the newly installed SDK */
-	sdkPath := filepath.Join(util.WorkspaceSdksDir, blob.Name)
+	sdkPath := filepath.Join(workspacebackend.WorkspaceSdksDir, blob.Name)
 
 	fs, err := m.backend.GetWorkspaceFs(ctx, workspace)
 	if err != nil {
@@ -286,7 +285,7 @@ func (m *SdkManager) undoLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
 
 			/* Update the workspace config */
 			err = m.backend.AddWorkspaceConfig(ctx, workspace,
-				&backend.WorkspaceConfigValue{
+				&workspacebackend.WorkspaceConfigValue{
 					Name:  "user.workspace.sdk",
 					Value: string(newSequence),
 				})
@@ -311,10 +310,10 @@ func (m *SdkManager) undoLinkSdk(task *state.Task, tomb *tomb.Tomb) error {
 		if newSeqLen > 0 {
 			/* There is another revision available, shift the link to it */
 			err = fs.Symlink(strconv.Itoa(int(sequence[blob.Name][newSeqLen-1].Revision)),
-				util.ToCurrentPath(blob.Name), true)
+				workspacebackend.SdkCurrentPath(blob.Name), true)
 		} else {
 			/* It was the only revision, remove the link */
-			err = fs.Remove(util.ToCurrentPath(blob.Name))
+			err = fs.Remove(workspacebackend.SdkCurrentPath(blob.Name))
 		}
 
 		return err
