+++ internal/workspacebackend/workspace_file_test.go
@@ -3,7 +3,6 @@ package workspacebackend_test
 import (
 	"testing"
 
-	util "github.com/canonical/workspace/internal"
 	"github.com/canonical/workspace/internal/workspacebackend"
 	"github.com/spf13/afero"
 	. "gopkg.in/check.v1"
@@ -31,7 +30,7 @@ sdks:
     channel: latest/stable
 `)
 	afero.WriteFile(f.fs, "/.workspace.xbert-gpu.yaml", buf, 0644)
-	file, err := workspacebackend.ReadWorkspace(f.fs, util.ToPathname("/", "xbert-gpu"))
+	file, err := workspacebackend.ReadWorkspace(f.fs, workspacebackend.WorkspaceFilePath("/", "xbert-gpu"))
 	c.Assert(err, Equals, nil)
 	c.Assert(file.Name, Equals, "xbert-gpu")
 	c.Assert(file.Base, Equals, "ubuntu@20.04")
